refactor: migrate helpers.options to typescript
