+++ src/helpers/types.ts
@@ -14,5 +14,6 @@ export * from './helpers.extras';
 export * from './helpers.interpolation';
 export * from './helpers.intl';
 export * from './helpers.math';
+export * from './helpers.options';
 export * from './helpers.rtl';
 export * from '../../types/helpers';
