+++ types/helpers/index.d.ts
@@ -1,4 +1,3 @@
 export * from './helpers.canvas';
-export * from './helpers.options';
 export * from './helpers.canvas';
 export * from './helpers.segment';
